Better error handling in init scripts

+++ my_init.d/99-ready.sh
@@ -12,4 +12,6 @@ echo "You can change the password using 'boxpasswd' command from shell"
 echo -e
 
 # Clear PASSWORD environment variable
-rm /etc/container_environment/PASSWORD
+if [ -f /etc/container_environment/PASSWORD ]; then
+	rm /etc/container_environment/PASSWORD
+fi
